feat: Make changes to the application data model

+++ app/controllers/api/v1/reservations_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::ReservationsController < ApplicationController
+  before_action :authenticate_user, only: [:create, :destroy]
   before_action :find_reservation, only: %i[show update destroy]
   ALLOWED_DATA = %i[check_in check_out user_id guests price hosting_id].freeze
 
@@ -15,18 +16,8 @@ class Api::V1::ReservationsController < ApplicationController
     begin
       @reservations = Reservation.where(user: params[:userId])
 
-      reservations = JSON.parse(@reservations.to_json({ include: [:hosting, { user: { except: :password_digest } }] }))
+      reservations = JSON.parse(@reservations.to_json({ include: [:hosting, { user: { except: :password_digest } }, { property: { include: [:images] }}] }))
       
-      reservations = reservations.map do |reservation|
-        property_id = reservation["hosting"]["property_id"]
-        property = Property.find(property_id)
-        images = PropertyImage.where(property:).limit(1)
-        property_hash = JSON.parse(property.to_json)
-        property_hash["images"] = images
-        reservation["property"] = property_hash
-        reservation
-      end
-
       render json: { success: true, data: reservations }, status: :ok
     rescue ActiveRecord::ActiveRecordError
       render json: { success: false, error: 'Internal server error.' }, status: :internal_server_error
@@ -44,21 +35,24 @@ class Api::V1::ReservationsController < ApplicationController
   end
 
   def create
-    @reservation = Reservation.new(create_params)
-    if @reservation.save
-      hosting = Hosting.find(params[:hosting_id])
-
-      @blocked_period = BlockedPeriod.new(property_id: hosting.property_id, start_date: params[:check_in],
+    hosting = Hosting.find(params[:hosting_id])
+    @blocked_period = BlockedPeriod.new(property_id: hosting.property_id, start_date: params[:check_in],
                                           end_date: params[:check_out])
-      if @blocked_period.save
+    if @blocked_period.save
+      @reservation = Reservation.new(create_params)
+      @reservation.property_id = hosting.property_id
+      @reservation.blocked_period_id = @blocked_period.id
+
+      if @reservation.save
         reservation = build_reservation(@reservation)
         render json: { success: true, data: reservation }, status: :created
       else
-        render json: { success: false, error: 'Cannot save reservation' }, status: 500
+        @blocked_period.destroy
+        render json: { success: false, error: @reservation.errors }, status: :unprocessable_entity
       end
     else
-      render json: { success: false, error: @reservation.errors }, status: :unprocessable_entity
-    end
+      render json: { success: false, error: @blocked_period.errors }, status: :bad_request
+    end 
   end
 
   def update
@@ -74,22 +68,25 @@ class Api::V1::ReservationsController < ApplicationController
   end
 
   def destroy
-    hosting = Hosting.find(@reservation.hosting_id)
-    reservation_blocked_period = BlockedPeriod.find_by(
-      property_id: hosting.property_id,
-      start_date: @reservation.check_in
-    )
-    reservation_blocked_period.destroy
-    @reservation.destroy
-    render json: { success: true, data: @reservation }, status: :ok
+    if @current_user.id == @reservation.user_id
+      @reservation.destroy
+      reservation_blocked_period = BlockedPeriod.find(@reservation.blocked_period_id)
+      reservation_blocked_period.destroy
+      
+      render json: { success: true, data: @reservation }, status: :ok
+    else
+      render json: { success: false, error: 'You are not authorized to complete this action.' }, status: :forbidden
+    end
   end
 
   private
 
   def find_reservation
-    @reservation = Reservation.find(params[:id])
-  rescue ActiveRecord::RecordNotFound
-    render json: { success: false, error: 'Reservation not found' }, status: :not_found
+    begin
+      @reservation = Reservation.find(params[:id])
+    rescue ActiveRecord::RecordNotFound
+      render json: { success: false, error: 'No reservation found with this ID.' }, status: :not_found
+    end
   end
 
   def create_params
@@ -97,13 +94,7 @@ class Api::V1::ReservationsController < ApplicationController
   end
 
   def build_reservation(reservation)
-    reservation_to_send = JSON.parse(reservation.to_json({ include: [:hosting, { user: { except: :password_digest } }] }))
-    property_id = reservation_to_send["hosting"]["property_id"]
-    property = Property.find(property_id)
-    images = PropertyImage.where(property:).limit(1)
-    property_hash = JSON.parse(property.to_json)
-    property_hash["images"] = images
-    reservation_to_send["property"] = property_hash
+    reservation_to_send = JSON.parse(reservation.to_json({ include: [:hosting, { user: { except: :password_digest } }, { property: { include: [:images] }}] }))
     reservation_to_send
   end
 end
